refactor flask app
