+++ api/route_q3.py
@@ -1,0 +1,7 @@
+
+from api.main import app
+from api.dbutils import conn
+
+@app.route('/api/test', methods=['GET'])
+def test():
+    return 'hello'
