+++ react-ui/Makefile
@@ -1,0 +1,3 @@
+
+run:
+	npm start
